MongoDB storage/retrieval integration.
Ability to store and retrieve scheduled duty calendars in the form of an
ArrayList<DutyBlock> to and from a MongoDB collection. Codecs for the
encoding/decoding of DutyBlock, Ra and LocalDate objects created, along
with associated CodecProviders. ScheduledDuty objects and associated
Codec/CodecProvider objects are used to store the ArrayList<DutyBlock>
within the database.